test/integration/lambda/requirements.txt
boto3
moto
aws-psycopg2==1.2.1
